refactor(server): sweep fragments, rename shared/ to channel/support/
- execution/worker-run-state.ts -> worker-runner-types.ts (3 src + 3 test
  importers repointed; audit counted 1)
- shared/sleep.ts -> shared/fetch-retry.ts (channel pollers repointed)
- context/assembler.ts + context/index.ts -> context/middleware.ts;
  bootstrap imports mcp-config directly, worker-runner imports middleware
  directly; barrel-shape integration test updated to the new wiring
- agents/runtime-definition.ts (single parse delegation) inlined at
  bootstrap/worker-bootstrap.ts with its egress-boundary comment; test
  relocated to test/bootstrap/runtime-agent-egress.test.ts
- src/shared/ -> src/channel/support/ (all importers are channel/*)
