+++ apps/server/src/channel/github/normalizer.ts
@@ -1,5 +1,5 @@
 import { Adapter } from "@openomni/protocol";
-import { normalizeContent } from "../../shared/trigger";
+import { normalizeContent } from "../support/trigger";
 import type { GitHubEventContent } from "./types";
 
 export interface GitHubNormalizerContext {
