+++ apps/server/src/channel/github/surface.ts
@@ -1,5 +1,5 @@
 import { Adapter, Operational, PolicyDecision } from "@openomni/protocol";
-import { Dedupe } from "../../shared/dedupe";
+import { Dedupe } from "../support/dedupe";
 import { GitHubClient } from "./client";
 import { GitHubNormalizer } from "./normalizer";
 import type { GitHubEventContent, GitHubIssueCommentPayload, GitHubIssuesPayload } from "./types";
