+++ apps/server/src/channel/telegram/client.ts
@@ -1,5 +1,5 @@
 import { Operational } from "@openomni/protocol";
-import { fetchWithRetry } from "../../shared/fetch-retry";
+import { fetchWithRetry } from "../support/fetch-retry";
 import type { ChannelClient, PublishPort } from "../types";
 import type { TelegramResponse, TelegramUser } from "./types";
