--- apps/server/src/context/index.ts
@@ -1,2 +1,0 @@
-export { McpConfigLoader } from "./mcp-config";
-export { createContextMiddleware } from "./middleware";
