--- apps/server/src/shared/sleep.ts
@@ -1,3 +1,0 @@
-export function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
