+++ apps/server/test/bootstrap/runtime-agent-egress.test.ts
@@ -1,9 +1,11 @@
 import { describe, expect, it } from "bun:test";
 import { WorkerBootstrap } from "@openomni/protocol";
-import { RuntimeAgentDefinition } from "../../src/agents/runtime-definition";
 import type { AgentDefinition, AgentPromptMetadata } from "../../src/agents/types";
 
-describe("RuntimeAgentDefinition", () => {
+// Pins the sole server-to-worker agent egress boundary used by
+// assembleBootstrap (bootstrap/worker-bootstrap.ts): protocol parsing
+// projects the wire contract and strips server-only prompt metadata.
+describe("runtime agent egress projection", () => {
   it("serializes server agent definitions through the protocol runtime schema", () => {
     const definition: AgentDefinition = {
       name: "dev",
@@ -18,7 +20,7 @@ describe("RuntimeAgentDefinition", () => {
       budget: { maxTurns: 4, maxToolCalls: 8 },
     };
 
-    const runtime = RuntimeAgentDefinition.create(definition);
+    const runtime = WorkerBootstrap.RuntimeAgentDefinition.parse(definition);
 
     expect(runtime).toEqual(definition);
     expect(WorkerBootstrap.RuntimeAgentDefinition.parse(runtime)).toEqual(runtime);
@@ -39,7 +41,7 @@ describe("RuntimeAgentDefinition", () => {
       metadata,
     } as AgentDefinition & { metadata: AgentPromptMetadata };
 
-    const runtime = RuntimeAgentDefinition.create(definitionWithMetadata);
+    const runtime = WorkerBootstrap.RuntimeAgentDefinition.parse(definitionWithMetadata);
 
     expect(runtime).toEqual({
       name: "dev",
