+++ apps/server/test/execution/worker-child-agent-notify.test.ts
@@ -3,7 +3,7 @@ import type { AgentResult } from "@openomni/agent";
 import { InjectionQueue } from "@openomni/openomni";
 import type { Tool } from "@openomni/protocol";
 import { Storage } from "@openomni/session";
-import type { WorkerRunState } from "../../src/execution/worker-run-state";
+import type { WorkerRunState } from "../../src/execution/worker-runner-types";
 import { WorkerRunner } from "../../src/execution/worker-runner";
 
 type SpawnRunOptions = Parameters<typeof WorkerRunner.spawnRun>[0];
